Tidied up code, working computer readable file for graphing, optimized code during testing, commented code. CSV file made for graphing.
