+++ testOutput.txt
@@ -1,58 +1,79 @@
+
+----------------------------------------
+
+HUMAN READABLE BENCHMARK RESULTS:
+
+
 BENCHMARK FOR QUICK SORT: 
 unsorted list: 
-Execution Time: 12ms.
+Execution Time: 11ms.
 unsorted list (reversed): 
-Execution Time: 13ms.
+Execution Time: 9ms.
 pre-sorted list: 
-Execution Time: 10ms.
+Execution Time: 3ms.
 
 Program Benchmark Results: 
-Average Execution Time: 11ms
+Average Execution Time: 7ms
 
 
 BENCHMARK FOR INSERTION SORT: 
 unsorted list: 
-Execution Time: 70ms.
+Execution Time: 25ms.
 unsorted list (reversed): 
-Execution Time: 19ms.
+Execution Time: 26ms.
 pre-sorted list: 
 Execution Time: 0ms.
+
+Program Benchmark Results: 
+Average Execution Time: 17ms
+
+
+BENCHMARK FOR SELECTION SORT: 
+unsorted list: 
+Execution Time: 44ms.
+unsorted list (reversed): 
+Execution Time: 44ms.
+pre-sorted list: 
+Execution Time: 1ms.
 
 Program Benchmark Results: 
 Average Execution Time: 29ms
 
 
-BENCHMARK FOR SELECTION SORT: 
-unsorted list: 
-Execution Time: 35ms.
-unsorted list (reversed): 
-Execution Time: 34ms.
-pre-sorted list: 
-Execution Time: 1ms.
-
-Program Benchmark Results: 
-Average Execution Time: 23ms
-
-
 BENCHMARK FOR HYBRID QUICK-INSERTION SORT: 
 unsorted list: 
-Execution Time: 6ms.
+Execution Time: 7ms.
 unsorted list (reversed): 
-Execution Time: 6ms.
+Execution Time: 9ms.
 pre-sorted list: 
 Execution Time: 2ms.
 
 Program Benchmark Results: 
-Average Execution Time: 4ms
+Average Execution Time: 6ms
 
 
 BENCHMARK FOR HYBRID QUICK-SELECTION SORT: 
 unsorted list: 
-Execution Time: 6ms.
+Execution Time: 8ms.
 unsorted list (reversed): 
-Execution Time: 6ms.
+Execution Time: 8ms.
 pre-sorted list: 
 Execution Time: 2ms.
 
 Program Benchmark Results: 
-Average Execution Time: 4ms
+Average Execution Time: 6ms
+
+----------------------------------------
+
+COMPUTER READABLE BENCHMARK RESULTS:
+
+
+title,normalsort,reversesort,sortedsort,averagetime
+quicksort,5,3,2,3
+insertionsort,15,15,0,10
+selectionsort,36,35,2,24
+hybridqisort,6,6,2,4
+hybridqssort,6,6,1,4
+
+----------------------------------------
+
